Add some unit tests, refactor wrapper
